requirements.txt
django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.6
stripe==8.4.0
python-dotenv==1.0.0
markdown==3.5.2
